# Improve performance and readability