--- a/.travis.yml
+++ b/.travis.yml
@@ -3,9 +3,6 @@ language: ruby
 sudo: false
 cache: bundler
 rvm:
-  - 2.3.0
-  - 2.3.8
-  - 2.4.5
   - 2.5.3
   - 2.6.3
   - ruby-head