--- a/benchmark/Gemfile
+++ b/benchmark/Gemfile
@@ -1,0 +1,5 @@
+source "https://rubygems.org"
+
+gem 'benchmark_driver'
+gem 'teyu', path: '../'
+gem 'attr_extras'