--- a/lib/teyu.rb
+++ b/lib/teyu.rb
@@ -16,43 +16,59 @@ module Teyu
 
     def apply
       sorter = Teyu::ArgsSorter.new(@arg_names)
+      # NOTE: accessing local vars is faster than method calls, so cache to local vars
+      required_positional_args = sorter.required_positional_args
+      required_keyword_args = sorter.required_keyword_args
+      optional_keyword_args = sorter.optional_keyword_args
+      keyword_args = sorter.keyword_args
 
       validate_number_of_required_positional_args = method(:validate_number_of_required_positional_args)
       validate_required_keyword_args = method(:validate_required_keyword_args)
+      validate_keyword_args = method(:validate_keyword_args)
 
       @klass.send(:define_method, :initialize) do |*arg_values|
-        validate_number_of_required_positional_args.call(sorter.required_positional_args, arg_values)
-        validate_required_keyword_args.call(sorter.required_keyword_args, arg_values)
+        if arg_values.last.is_a?(Hash)
+          positional_arg_values = arg_values[0...-1]
+          keyword_arg_values = arg_values.last
+        else
+          positional_arg_values = arg_values
+          keyword_arg_values = {}
+        end
 
-        sorter.required_positional_args.zip(arg_values).each do |name, value|
+        validate_number_of_required_positional_args.call(required_positional_args, positional_arg_values)
+        validate_required_keyword_args.call(required_keyword_args, keyword_arg_values)
+        validate_keyword_args.call(keyword_args, keyword_arg_values)
+
+        required_positional_args.zip(arg_values).each do |name, value|
           instance_variable_set(:"@#{name}", value)
         end
 
-        sorter.optional_keyword_args.each do |name, value|
+        optional_keyword_args.each do |name, value|
           instance_variable_set(:"@#{name}", value)
         end
 
-        keyword_arg_values = arg_values[sorter.required_positional_args.length..].find { |value| value.is_a?(Hash) } || {}
-        (sorter.keyword_args & keyword_arg_values.keys).each do |name|
-          instance_variable_set(:"@#{name}", keyword_arg_values[name])
+        keyword_arg_values.each do |name, value|
+          instance_variable_set(:"@#{name}", value)
         end
       end
     end
 
     private
 
-    def validate_number_of_required_positional_args(arg_names, arg_values)
-      req_arg_names_count = arg_names.count
-      req_arg_values_count = arg_values.filter { |value| !value.is_a?(Hash) }.count
-
-      raise ArgumentError, "wrong number of arguments (given #{req_arg_values_count}, expected #{req_arg_names_count})" if req_arg_names_count != req_arg_values_count
+    def validate_number_of_required_positional_args(required_positional_args, positional_arg_values)
+      if required_positional_args.count != positional_arg_values.count
+        raise ArgumentError, "wrong number of arguments (given #{positional_arg_values.count}, expected #{required_positional_args.count})"
+      end
     end
 
-    def validate_required_keyword_args(arg_names, arg_values)
-      arg_names.each do |name|
-        keyword_arg_keys = arg_values.find { |value| value.is_a?(Hash) }&.keys || []
-        raise ArgumentError, "(missing keyword: #{name})" unless keyword_arg_keys.include?(name)
-      end
+    def validate_required_keyword_args(required_keyword_args, keyword_arg_values)
+      missing_keywords = required_keyword_args - keyword_arg_values.keys
+      raise ArgumentError, "missing keywords: #{missing_keywords.join(', ')}" unless missing_keywords.empty?
+    end
+
+    def validate_keyword_args(keyword_args, keyword_arg_values)
+      unknown_keywords = keyword_arg_values.keys - keyword_args
+      raise ArgumentError, "unknown keywords: #{unknown_keywords.join(', ')}" unless unknown_keywords.empty?
     end
   end