--- a/test/args_sorter_test.rb
+++ b/test/args_sorter_test.rb
@@ -5,20 +5,20 @@ class Teyu::ArgsSorterTest < Test::Unit::TestCase
     @sorter = Teyu::ArgsSorter.new([:a, :b, :c!, { d: 'd' }, { e: 'e' }, :f!])
   end
 
-  def test_req_args
-    assert { @sorter.req_args == [:a, :b] }
+  def test_required_positional_args
+    assert { @sorter.required_positional_args == [:a, :b] }
   end
 
   def test_keyword_args
     assert { @sorter.keyword_args == [:c, :f, :d, :e] }
   end
 
-  def test_keyreq_args
-    assert { @sorter.keyreq_args == [:c, :f] }
+  def test_required_keyword_args
+    assert { @sorter.required_keyword_args == [:c, :f] }
   end
 
-  def test_key_args
-    assert { @sorter.key_args == { d: 'd', e: 'e' } }
+  def test_optional_keyword_args
+    assert { @sorter.optional_keyword_args == { d: 'd', e: 'e' } }
   end
 end