--- a/test/argument_test.rb
+++ b/test/argument_test.rb
@@ -1,0 +1,24 @@
+require_relative "test_helper"
+
+class Teyu::ArgumentTest < Test::Unit::TestCase
+  def setup
+    @argument = Teyu::Argument.new([:a, :b, :c!, { d: 'd' }, { e: 'e' }, :f!])
+  end
+
+  def test_required_positional_args
+    assert { @argument.required_positional_args == [:a, :b] }
+  end
+
+  def test_keyword_args
+    assert { @argument.keyword_args == [:c, :f, :d, :e] }
+  end
+
+  def test_required_keyword_args
+    assert { @argument.required_keyword_args == [:c, :f] }
+  end
+
+  def test_optional_keyword_args
+    assert { @argument.optional_keyword_args == { d: 'd', e: 'e' } }
+  end
+end
+