* Add benchmark script
* Add unknown keywords validation
* Fix travis to work
* And, improve performance and readability

Before:

```
Warming up --------------------------------------
              Normal     1.180M i/s -      1.235M times in 1.046608s (847.32ns/i)
                teyu   125.081k i/s -    130.812k times in 1.045816s (7.99μs/i)
         attr_extras   226.015k i/s -    232.401k times in 1.028254s (4.42μs/i)
Calculating -------------------------------------
              Normal     1.353M i/s -      3.541M times in 2.616627s (739.04ns/i)
                teyu   153.047k i/s -    375.243k times in 2.451815s (6.53μs/i)
         attr_extras   217.607k i/s -    678.045k times in 3.115912s (4.60μs/i)

Comparison:
              Normal:   1353104.2 i/s
         attr_extras:    217607.2 i/s - 6.22x  slower
                teyu:    153047.0 i/s - 8.84x  slower
```

After:

```
Warming up --------------------------------------
              Normal     1.305M i/s -      1.376M times in 1.054138s (766.26ns/i)
                teyu   341.528k i/s -    343.680k times in 1.006302s (2.93μs/i)
         attr_extras   217.065k i/s -    225.348k times in 1.038158s (4.61μs/i)
Calculating -------------------------------------
              Normal     1.273M i/s -      3.915M times in 3.074409s (785.26ns/i)
                teyu   335.199k i/s -      1.025M times in 3.056641s (2.98μs/i)
         attr_extras   208.963k i/s -    651.195k times in 3.116313s (4.79μs/i)

Comparison:
              Normal:   1273457.8 i/s
                teyu:    335199.0 i/s - 3.80x  slower
         attr_extras:    208963.3 i/s - 6.09x  slower
```

About 2 times faster than before.